+++ GoodBuyUS/src/main/java/dev_final_team10/GoodBuyUS/domain/user/entity/User.java
@@ -1,6 +1,6 @@
-package dev_final_team10.GoodBuyUS.domain.user;
+package dev_final_team10.GoodBuyUS.domain.user.entity;
 
-import dev_final_team10.GoodBuyUS.domain.neighborhood.Neighborhood;
+import dev_final_team10.GoodBuyUS.domain.neighborhood.entity.Neighborhood;
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.security.crypto.password.PasswordEncoder;
