+++ GoodBuyUS/src/main/java/dev_final_team10/GoodBuyUS/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package dev_final_team10.GoodBuyUS.repository;
 
-import dev_final_team10.GoodBuyUS.domain.user.User;
+import dev_final_team10.GoodBuyUS.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
