+++ GoodBuyUS/src/main/java/dev_final_team10/GoodBuyUS/service/ExcelService.java
@@ -1,6 +1,6 @@
 package dev_final_team10.GoodBuyUS.service;
 
-import dev_final_team10.GoodBuyUS.domain.neighborhood.Neighborhood;
+import dev_final_team10.GoodBuyUS.domain.neighborhood.entity.Neighborhood;
 import dev_final_team10.GoodBuyUS.repository.NeighborhoodRepository;
 import lombok.RequiredArgsConstructor;
 import org.apache.poi.ss.usermodel.Row;
